JSON::Pointer + JSON::Patch: Fix for large objects and patches
https://github.com/zigorou/perl-json-patch/issues/3

The JSON::Pointer module used by JSON::Patch is badly designed for
JSON::Patch: it creates a copy of the document for each operation in a
patch. This can become very slow for large objects.

In a real-world example, applying a 1661-operation patch to a large
object takes 110 seconds for me, spending > 80% in Clone::clone. It is
way faster to compute the 1661-operation patch between the two large
objects than to apply it.

When JSON::Pointer is changed to modify the document in place and only a
single copy is created in JSON::Patch::patch, the runtime drops to 0.3
seconds.

This fix changes JSON::Pointer in an incompatible way.  It is unlikely
that upstream will adopt this fix, if those modules are even still
maintained.  The upstream repsitories are:

  git://github.com/zigorou/perl-json-pointer.git
  git://github.com/zigorou/perl-json-patch.git