Fix rider shown on results pages
On the client-side results page, only show the actual starters (i.e.,
only registered riders if features.registered is set).  On the
server-side results page, show all verified, pre-registered riders as
long as there are no actual results.  Once there are actual results to
show, only show the actual starters there as well.

This is a bit of a hack, but it works around the problem that we don't
have a proper view that shows pre-registered riders (and possible which
days they are registered for) so far.