Sync: Fix starting sync when not logged in on target
When not logged in on the sync target, we redirect to the target's login
page.  When the login there succeeds, the target redirects back to us,
with an additional sync parameter in the URL.  When we see that
parameter, we know that we have tried to synchronize, and we start that
sync.