Implement a very rudimentary kiosk mode
When the kiosk flag of a user in the database is set, online
registrations by this user will be done in kiosk mode.  This allows to
add new riders and to set their email address, but doesn't allow to
modify existing riders.  Whenever a new rider is created, an alert
dialog tells the operator which number (a negative number) the user was
assigned in the database so that the registration office can quickly
look up that rider for finishing the registration.