Fix bugs when searching for riders
When a single rider was found, the previous search result wasn't cleared
anymore due to not clearing the search result anymore when picking a
rider from the search result.

When a search returns no riders, move away from the previous rider so
that the previous rider isn't accidentally assumed to be the one the
user was looking for.