Update riders.user_tag based on users.user_tag
When a rider's email address is modified and that rider's user_tag is
still NULL, set the rider's user_tag to the user's user_tag so that the
user will have access to the rider data.  This happens both when a rider
is added or modified in the admin interface and when an event is
imported / synchronized between servers.

(This will cause some unnecessary differences when synchronizing. We
can later reduce the amount of differences by back-merging user tags
from the synchronization target to the source.)

When a user for a particluar email address is created, update user_tag
of all riders with that email address and no current user_tag to that
user's user_tag so that this user will immediately see all the available
riders.