admin: Fix inheritance of "registered" flags
When the "start_tomorrow" feature was enabled, creating a folow-up event
did inherit the "registered" flag of all riders.  This was causing
problems when activating the online registration because registered
riders can no longer set their "start" and "start_tomorrow" flags.  Fix
this by only inheriting the "registered" flag when the predecessor event
has the "start_tomorrow" feature enabled and both the rider's
"registered" and "start_tomorrow" flags are set.