Disable express etag support for dynamic results
By default, express uses etags for all requests, which causes the server
to respond with "304 Not Modified" when the data hasn't changed.  This
works well except for pages that require CORS headers: in that case, the
browser expects to see the appropriate CORS headers but the server only
responds with "304 Not Modified" without any CORS headers; as a
consequence, the browser discards the reply and causes the application
to fail.

Try fixing this by disabling etags for dynamically generated results
(but turn them on explicitly for the static pages).