admin: Compute series scores in Javascript
Convert the Perl code for computing the series scores to Javascript.  Store the
computed result in the database instead of always computing it on the fly.
Once we have a way to access the results pages in the admin interface from the
public, all the non-trivial pieces of Perl code still around can finally be
removed.