Add support for split rankings
Normally, a rider that takes part in an event shows up in the overall ranking
of the class the rider is in.  Riders can be put in additional rankings, which
means that they will show up in those additional rankings as well.  When a
rider is put in a split ranking, it is excluded from the overall ranking.