admin: Allow to configure the main ranking
When an event has more than one non-split ranking, either of those rankings can
be shown in combination with the overall result of the event.  So far, which of
those rankings is chosen is configured in the settings of the results page.
In the future, this will be configured in the event settings, so add an option
for that.