admin: Fix Content-Disposition headers
When a Content-Disposition header specified that a file is "attachment",
Firefox will always open a dislog asking what to do with a file.  This is very
annoying for PDF files, for which it makes sense to open them automatically
either in the browser or in an external viewer.  Change those headers for PDF
files to "inline".  In addition, consistently URI-encode the filenames, even
for inline files so that Firefox wiill know what filename to use in case the
user wants to save the file.