More flexible PDF form handling
Allow to provide multiple registration forms per event type, and to serve
specific versions of a form depending on the properties of a rider:

Each *.pdf file in the pdf/regform/<event_type>/ directory is offered as a
fill-out form in events of that particular type.  When the filename of a form
contains expressions of the form {property:value}, this form is used only for
riders with a matching property.  When there are multiple similar filenames,
the matching the rider most specifically is chosen.  For example, with the
following forms available:

  pdf/regform/otsv-acup/Registration Form.pdf
  pdf/regform/otsv-acup/Number Plate.pdf
  pdf/regform/otsv-acup/Number Plate{color:008000}.pdf

The event type "otsv-acup" has two forms available.  When creating a number
plate, file "Number Plate{color:008000}.pdf is chosen for riders in a "green"
class (008000 is a RGB value that corresponds to green), and "Number Plate.pdf"
is chosen for all other riders.