Fix "events so far" for joint rankings
When a ranking requires a minimum number of events to participate
in, we compute the number of events that have taken place so far
in order to determine if a rider can still participate in enough
events to make it into the ranking; we only "exclude" riders from
the ranking once they cannot participate in enough events anymore.

This computation was broken for joint rankings; fix that.