event results: Fix row height for rounds not ridden yet
In the event results page, when showing the individual marks of each zone,
output zero-width space characters when in the marks-per-round column when
there are no marks for that round, yet.  This ensures that rounds that haven't
been started yet will be shown with the same row height as all the other
rounds.