Make compute_event module more reusable (2)
Pass in and return a list of riders instead of a hash indexed by the
rider number.  In the future, with combined rankings across multiple
events, the same rider could start in different classes in different
events, and then the rider number won't be unique anymore.

As a side effect, this simplifies the code.