angular: There is no conditional ng-repeat
In tags that contain ng-repeat and ng-if, the ng-if is evaluated for each
repeated item no matter which order the ng-repeat and ng-if tags are in.  There
is way to exclude the tag entirely if a condition isn't met.  This can cause a
significant overhead.

Work around that by changing <foo ng-if="condition" ng-repeat="item in items">
tags to <foo ng-repeat=item in (condition ? items : [])">.