Combined rankings: Leave event result blank for riders who are not starting
In combined rankings, riders may not start in all of the events
comprised by the ranking.  In that case, instead of leaving all fields
of the result blank which would indicate that no results have been
enetered, show a single, empty table cell.  I hope this won't be too
visually distracting.