Truncate failure text when necessary
When there is not enough space in a results table for the full reason of a
failure, truncate the reason instead of expanding the table columns.  This
would be quite noticable in A-Cup events because those events have less space
for the reason of a failure, but in these events, instead of marking riders as
failed, we pretent that those riders have failed all sections they didn't ride.