pdf-fill-form.py: Support multiple forms per page
Allow to pass multiple pdf files to the script, each with a different number of
forms per document.  Each field name ends in "%" followed by the form number,
so in a document with three forms, there would be three "name" fields called
"name%1", "name%2", and "name%3", for example.

With that, the script chooses a combination of documents it fills out such that
no forms remain blank (if possible) and the number of documents (and hence
pages) is minimized.

Maybe we'll not ever use that.