Add country and hide_country settings
Define the country an event is taking place in and allow to hide that country
from results and events lists.  So far, we have always assumed that events are
in Austria and that Austria should be hidden on events lists, but that may not
be true for all events anymore this year.