Allow to enter results section by section
Add an option to enter results section by section instead of round by round:
usually, results are collected on a score card that travels with the riders and
records the marks of all zones in a round.  The score cards are switched in the
race office, where the results for previous rounds can then be entered.  During
data entry, we switch from zone to zone.

When the results are collected on per-zone score sheets, all the results arrive
at the race office after all riders have completed the last round of that zone,
and all rounds of that zone are entered at once.  In that case, we switch from
round to round before moving to the next zone.  In addition, riders are sorted
by number on those score sheets, so we can automatically jump from rider to
rider during data entry.