marks form: Improve ArrowLeft / ArrowRight handling
When ArrowLeft is pressed, jump to the previous zone when the entire marks
field is selected or the cursor is at the beginning of the field.  Likewise,
when ArrowRight is pressed, jump to the next zone when the entire marks field
is selected or the cursor is at the end of the field.