marks form: When entering results per zone, stay in that zone
So far, when loading a new rider, we were jumping to the first zone that hasn't
been entered, yet.  When entering data zone-by-zone, that could be a previous
zone that hasn't been entered completely yet.  This is not desirable, so change
that to stay in the current zone instead.  (This doesn't affect round-by-round
data entry.)