Add support for direct printing
Reorganize the pdf forms to be organized in form types (for example, forms and
labels).  Add a "printers" hash to backend/config.json that allows to define
a default printer for each form type.

When a default printer is defined for a particular form type, a button with a
printer icon is added next to the form name in the admin interface.  Pressing
that button will print the pdf form to the configured printer directly instead
of opening the pdf document.  This will make printing name labels much faster.