Allow to change the direction of entry in the marks form
Add a small button in the left/top corner of the marks table that indicates the
current direction of marks entry (round wise vs. zone wise) and allows to flip
the direction.  This is better than configuring the direction in the settings
for events that use both methods depending on the zone.