Adapt marks computation to mobile scoring
So far, rider scores were mostly entered round by round, and zones skipped in a
round could immediately be marked as skipped.  When computing the results of an
event, results were only counted up to the first zone that hasn't been entered
yet.

With mobile scoring, the results from different zones can arrive in a very
different order, so it makes more sense to count all the results.  Riders are
now ranked by the number of completed rounds, the number of completed zones,
and only then by their marks, etc.  The results table indicates which rounds
are still incomplete.