Rewrite sync code
When synchronizing an event from host A to host B, also transfer the
scoring data between A and B: while the non-scoring data is only
transferred from A to B, the scoring data is transferred in both
directions.  Each side computes the event results based on the scoring
data it already has.  The scoring data will eventually converge either
because the scoring devices manage to talk to host A as well as B or
through the sync.  At that point, the event results on both sides will
be identical.

With the old promise-based sync code, even the relatively simple sync
code was difficult to maintain.  Since browsers now widely support
async/await, use that instead.  The result is much better now even
though the code could be cleaned up further.