Make corrected marks work in the backend
We compute which round a rider is in by bringing the marks of the rider
into chronological order.  When a result is corrected, the result counts
at the time it was initially entered, not at the time when it was
corrected, so we need to be careful.

This is different from when a result is canceled ("deleted") and a new
result is entered instead: in that case, the new result counts at the
time it was entered.