Split rankings fix
When a rider is both in the ranking that is shown together with the
overall results (event.main_ranking) and in a ranking that splits riders
from the overall results, include the rider in the overall results.
That way, it will be obvious which riders should be taken out of the
main ranking.