Add event.ctime timestamp
event.mtime indicates the last time the event was last modified, either
because the event itself was changed, because a rider has changed, or
because a scoring update has occurred.

event.ctime indicates the last time the event was last modified, either
because the event itself was changed or because a rider has changed,
but not because a scoring update has occurred.

Scoring updates can end up modifying multiple copies of an event
independently (i.e., the source and target of a sync), each affecting
the mtime, so the mtime cannot be used for determining which copy has
the more recent data.  On the other hand, the ctime can be used for
that.

Include the ctime in GET/api/scoring/:access_token replies.