Fix error showing event results
In some situations, the browser can fail to show the results of an event
because marks_per_zone[round] is undefined.  This is because some
compensating code earlier in the same function fails.  This
compensating code still needs fixing, but for now, just handle undefined
marks_per_zone[round] values here as well.