Fix PageUp / PageDown keys in search fields
So far, all keydown events in search fields were disabled, which
allows the onchange action to trigger but also prevents reacting to
PageUp and PageDown events.  Change that to only disable handlng of the
Enter key for input fields that have an onchange attribute.  By checking
for onchange attributes, we can also eliminate class 'default-keydown'.