Get pdf conversion working again with weasyprint 51
On RHEL (running on the server side), we cannot upgrade to the latest
version of WeasyPrint because the python3 and pango packages are too
old.  However, weasyprint 51 broke when the "-t pdf" option was removed;
old versions of weasyprint expect "-t pdf" to be specified when the file
name doesn't end in ".pdf".  Make sure the temp file names end in ".pdf"
to make old as well as new versions of weasyprint happy.