Fix email comparison when registering riders
To give users access to their own data via the registration form even if
those data have originally been entered in the race office ("Nennbüro"),
the registration form lists riders registered under the same user_tag as
well as the same email address.  The comparison operation for e-mail
addresses is case insensitive.

However, when saving the registration form, the backend checks if the
user tag or the email matches, but using a case sensitive comaprison
operation.  So if the email address is spelled differently, users can
get access to a rider record which they then cannot save.

Fix that by using the same comparison operation in both places.