Flip how "no_ranking1" is shown in the event settings
Instead of checking "No ranking 1" to take all riders in a class out of
the first ranking, the event settings now have a "Ranking 1" checkbox
which is checked by default, and which needs to be unchecked to achieve
the same effect.  (In the code, this is still implemented by the
"no_ranking1" flag.)

Instead of changing the model value for this checkbox, we add an
'inverted' directive that transparently flips the model's value.

To get long ranking descriptions to hyphenate, define the overall
language and set the hyphens css attribute of ranking 1's column
heading.