Make drop events work more intuitively
In event series, for each class in the ranking, a number of results to
drop from the ranking can be defined ("Streichresultate").

For example, when there are 10 events and 2 drop results, a maximum of
10 - 2 = 8 events will considered in the ranking.  For riders who
participated in N events with N > 8, their worst N - 8 results will be
ignored.

In addition, the total number of events can be defined as well.  This is
useful information for participants, but it also allows to define the
number of drop events independent from how many events have already
taken place.

So far, drop events were only considered when the total number of events
was defined as well.  This is confusing, so change that to also take the
drop events into account when the total number of events is not defined.