Fix sections list view for empty classes
The section lists include which class each rider is registered in and
ranking class each rider is in to allow filtering by class, but the list
of classes only contained the list of registered classes.  This will
cause classes that don't have and riders directly riding in to be
missing, so the section lists view will fail when trying to sort by
ranking class.  Fix that by making sure all ranking classes are
included as well.