Fix results table column alignment
Fix the column alignment in the results table for riders that have
failed if their ranking class requires tie breaking.

(In rankings that span multiple events, it's arguably not very useful to
have a tie break column per event; it would be more useful to calculate
and display tie breaking based on only the last event.  But this would
create even more problems for riders that didn't participate in the last
event ...)